+++ app/v0/cipal/types/msgs.go
@@ -32,11 +32,11 @@ type ADParam struct {
 
 type CIPALUserRequest struct {
 	Params ADParam           `json:"params" yaml:"params"`
-	Sig    auth.StdSignature `json:"signature" yaml:"signature`
+	Sig    auth.StdSignature `json:"signature" yaml:"signature"`
 }
 
 type MsgCIPALClaim struct {
-	From        sdk.AccAddress   `json:"from" yaml:"from`
+	From        sdk.AccAddress   `json:"from" yaml:"from"`
 	UserRequest CIPALUserRequest `json:"user_request" yaml:"user_request"`
 }