+++ app/v0/ipal/client/rest/rest.go
@@ -1,10 +1,12 @@
 package rest
 
 import (
-	"github.com/netcloth/netcloth-chain/client/context"
 	"github.com/gorilla/mux"
+
+	"github.com/netcloth/netcloth-chain/client/context"
 )
 
+// RegisterRoutes registers the routes from the different modules for the LCD.
 func RegisterRoutes(cliCtx context.CLIContext, r *mux.Router) {
 	registerQueryRoutes(cliCtx, r)
 }