+++ app/v0/mock/p0/protocol_v0.go
@@ -34,7 +34,7 @@ import (
 
 var _ protocol.Protocol = (*ProtocolV0)(nil)
 
-// The module BasicManager is in charge of setting up basic,
+// ModuleBasics is in charge of setting up basic,
 // non-dependant module elements, such as codec registration
 // and genesis verification.
 var ModuleBasics = module.NewBasicManager(
@@ -252,6 +252,7 @@ func (p *ProtocolV0) configKeepers() {
 		p.Cdc,
 		protocol.Keys[protocol.VmStoreKey],
 		protocol.Keys[protocol.VmCodeStoreKey],
+		protocol.Keys[protocol.VmLogStoreKey],
 		protocol.Keys[protocol.VmStoreKey],
 		vmSubspace,
 		p.AccountKeeper)