+++ app/v0/vm/logger_test.go
@@ -47,6 +47,7 @@ func TestStoreCapture(t *testing.T) {
 		types.ModuleCdc,
 		sdk.NewKVStoreKey(StoreKey),
 		sdk.NewKVStoreKey(CodeKey),
+		sdk.NewKVStoreKey(LogKey),
 		sdk.NewKVStoreKey(StoreDebugKey),
 		paramsKeeper.Subspace(bank.DefaultParamspace),
 		accountKeeper)