+++ app/v0/vm/types/statedb.go
@@ -1,18 +1,18 @@
 package types
 
 import (
+	"encoding/json"
 	"fmt"
 	"math/big"
 	"os"
 	"sort"
 	"sync"
 
-	ethstate "github.com/ethereum/go-ethereum/core/state"
-
 	"github.com/tendermint/tendermint/crypto"
 
 	"github.com/netcloth/netcloth-chain/app/v0/auth"
 	"github.com/netcloth/netcloth-chain/app/v0/auth/types"
+	"github.com/netcloth/netcloth-chain/app/v0/vm/common/math"
 	sdk "github.com/netcloth/netcloth-chain/types"
 )
 
@@ -34,6 +34,7 @@ type CommitStateDB struct {
 	ak              auth.AccountKeeper
 	storageKey      sdk.StoreKey
 	codeKey         sdk.StoreKey
+	logKey          sdk.StoreKey
 	storageDebugKey sdk.StoreKey
 
 	// maps that hold 'live' objects, which will get modified while processing a
@@ -47,7 +48,6 @@ type CommitStateDB struct {
 	thash, bhash sdk.Hash
 	txIndex      int
 	logs         map[sdk.Hash][]*Log
-	logSize      uint
 
 	// TODO: Determine if we actually need this as we do not need preimages in
 	// the SDK, but it seems to be used elsewhere in Geth.
@@ -78,11 +78,12 @@ type CommitStateDB struct {
 // CONTRACT: Stores used for state must be cache-wrapped as the ordering of the
 // key/value space matters in determining the merkle root.
 //func NewCommitStateDB(ctx sdk.Context, ak auth.AccountKeeper, storageKey, codeKey sdk.StoreKey) *CommitStateDB {
-func NewCommitStateDB(ak auth.AccountKeeper, storageKey, codeKey, storageDebugKey sdk.StoreKey) *CommitStateDB {
+func NewCommitStateDB(ak auth.AccountKeeper, storageKey, codeKey, logKey, storageDebugKey sdk.StoreKey) *CommitStateDB {
 	return &CommitStateDB{
 		ak:                ak,
 		storageKey:        storageKey,
 		codeKey:           codeKey,
+		logKey:            logKey,
 		storageDebugKey:   storageDebugKey,
 		stateObjects:      make(map[string]*stateObject),
 		stateObjectsDirty: make(map[string]struct{}),
@@ -178,9 +179,8 @@ func (csdb *CommitStateDB) AddLog(log *Log) {
 	log.TxHash = csdb.thash
 	log.BlockHash = csdb.bhash
 	log.TxIndex = uint(csdb.txIndex)
-	log.Index = csdb.logSize
+	log.Index = csdb.updateLogIndexByOne(false)
 	csdb.logs[csdb.thash] = append(csdb.logs[csdb.thash], log)
-	csdb.logSize++
 }
 
 // AddPreimage records a SHA3 preimage seen by the VM.
@@ -308,12 +308,26 @@ func (csdb *CommitStateDB) GetCommittedState(addr sdk.AccAddress, hash sdk.Hash)
 }
 
 // GetLogs returns the current logs for a given hash in the state.
-func (csdb *CommitStateDB) GetLogs(hash sdk.Hash) []*Log {
-	return csdb.logs[hash]
+func (csdb *CommitStateDB) GetLogs(hash sdk.Hash) (logs []*Log) {
+	r, ok := csdb.logs[hash]
+	if ok {
+		return r
+	}
+
+	ctx := csdb.ctx
+	store := ctx.KVStore(csdb.logKey)
+	d := store.Get(hash.Bytes())
+	err := json.Unmarshal(d, &logs)
+	if err != nil {
+		ctx.Logger().Error(err.Error())
+		return
+	}
+
+	return
 }
 
 // Logs returns all the current logs in the state.
-func (csdb *CommitStateDB) Logs() []*Log {
+func (csdb *CommitStateDB) Logs() []*Log { // todo: is should get all logs from store?
 	var logs []*Log
 	for _, lgs := range csdb.logs {
 		logs = append(logs, lgs...)
@@ -323,7 +337,7 @@ func (csdb *CommitStateDB) Logs() []*Log {
 }
 
 func (csdb *CommitStateDB) ClearLogs() {
-	for k, _ := range csdb.logs {
+	for k := range csdb.logs {
 		delete(csdb.logs, k)
 	}
 }
@@ -390,6 +404,58 @@ func (csdb *CommitStateDB) Commit(deleteEmptyObjects bool) (root sdk.Hash, err e
 	return
 }
 
+func (csdb *CommitStateDB) commitLogs() {
+	ctx := csdb.ctx
+	store := ctx.KVStore(csdb.logKey)
+
+	var hs []string
+	for h := range csdb.logs {
+		hs = append(hs, h.String())
+	}
+	sort.Strings(hs)
+
+	for _, h := range hs {
+		hash := sdk.HexToHash(h)
+		d, err := json.Marshal(csdb.logs[hash])
+		if err != nil {
+			ctx.Logger().Error(err.Error())
+			continue
+		}
+
+		ctx.Logger().Debug("save log----", hash.String(), ":", string(d))
+		store.Set(hash.Bytes(), d)
+	}
+}
+
+func (csdb *CommitStateDB) updateLogIndexByOne(isSubtract bool) uint64 {
+	ctx := csdb.ctx
+	store := ctx.KVStore(csdb.logKey)
+
+	value := big.NewInt(0)
+	if store.Has(LogIndexKey) {
+		d := store.Get(LogIndexKey)
+		value.SetBytes(d)
+
+		if isSubtract {
+			if value.Uint64() == 0 {
+				ctx.Logger().Error(fmt.Sprintf("current logIndex is 0, can not to be Subtracted"))
+				return 0
+			}
+			value.SetUint64(value.Uint64() - 1)
+		} else {
+			if value.Uint64() == math.MaxUint64 {
+				ctx.Logger().Error(fmt.Sprintf("current logIndex will out of range"))
+				return value.Uint64()
+			}
+			value.SetUint64(value.Uint64() + 1)
+		}
+	}
+
+	store.Set(LogIndexKey, value.Bytes())
+
+	return value.Uint64()
+}
+
 // Finalise finalizes the state objects (accounts) state by setting their state,
 // removing the csdb destructed objects and clearing the journal as well as the
 // refunds.
@@ -420,6 +486,9 @@ func (csdb *CommitStateDB) Finalise(deleteEmptyObjects bool) {
 
 		csdb.stateObjectsDirty[addr] = struct{}{}
 	}
+
+	csdb.commitLogs()
+	csdb.ClearLogs()
 
 	// invalidate journal because reverting across transactions is not allowed
 	csdb.clearJournalAndRefund()
@@ -493,9 +562,9 @@ func (csdb *CommitStateDB) RevertToSnapshot(revID int) {
 
 // Database retrieves the low level database supporting the lower level trie
 // ops. It is not used in Ethermint, so it returns nil.
-func (csdb *CommitStateDB) Database() ethstate.Database {
-	return nil
-}
+//func (csdb *CommitStateDB) Database() ethstate.Database {
+//	return nil
+//}
 
 // Empty returns whether the state object is either non-existent or empty
 // according to the EIP161 specification (balance = nonce = code = 0).
@@ -547,7 +616,6 @@ func (csdb *CommitStateDB) Reset(_ sdk.Hash) error {
 	csdb.bhash = sdk.Hash{}
 	csdb.txIndex = 0
 	csdb.logs = make(map[sdk.Hash][]*Log)
-	csdb.logSize = 0
 	csdb.preimages = make(map[sdk.Hash][]byte)
 
 	csdb.clearJournalAndRefund()
@@ -627,7 +695,6 @@ func (csdb *CommitStateDB) Copy() *CommitStateDB {
 		stateObjectsDirty: make(map[string]struct{}, len(csdb.journal.dirties)),
 		refund:            csdb.refund,
 		logs:              make(map[sdk.Hash][]*Log, len(csdb.logs)),
-		logSize:           csdb.logSize,
 		preimages:         make(map[sdk.Hash][]byte),
 		journal:           newJournal(),
 	}