+++ tests/scripts/gen_account.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+bash batch_add_account.sh acc 1000 |grep "\"address\"" |awk -F '"' '{print $4}' >n1_accs