+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [unreleased]
+
+### nchd
+
+* fix upgrade module
+* fix out of memory caused by vm logs
+* fix consensus failure caused by vm
+* add and fix vm instruction test
+
 ## testnet-v1.2.0
 
 ### nchd